requirements.txt
# Core ML & Data Science
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
xgboost>=1.6.0
lightgbm>=3.3.0
scipy>=1.9.0

# Database Connectivity
mysql-connector-python>=8.0.0
SQLAlchemy>=1.4.0
pymysql>=1.0.2

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
kaleido>=0.2.1

# Data Processing & Utilities
joblib>=1.1.0
python-dotenv>=0.19.0
tqdm>=4.64.0

# Web Framework (for API endpoints)
fastapi>=0.85.0
uvicorn>=0.18.0
pydantic>=1.10.0

# Caching & Performance
redis>=4.3.0

# Jupyter & Development
jupyter>=1.0.0
notebook>=6.4.0
ipywidgets>=7.7.0

# Testing
pytest>=7.1.0
pytest-cov>=4.0.0

# Code Quality
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0

# Development Tools
pre-commit>=2.20.0